+++ src/test/java/io/github/zeroone3010/yahueapi/StateTest.java
@@ -167,7 +167,7 @@ class StateTest {
 
   @Test
   void testEffectConstants() {
-    assertEquals(State.COLOR_LOOP, State.builder().effect(EffectType.COLOR_LOOP));
+    assertEquals(State.COLOR_LOOP_EFFECT, State.builder().effect(EffectType.COLOR_LOOP));
     assertEquals(State.NO_EFFECTS, State.builder().effect(EffectType.NONE));
   }